Reorganized files so that the view and the languageSelector fragment are in separate packages
